+++ packages/appcd-default-plugins/CHANGELOG.md
@@ -1,5 +1,6 @@
-# v3.0.1
+# v4.0.0
 
+ * BREAKING CHANGE: Default plugins are now installed at runtime instead of postinstall.
  * chore: Updated dependencies
 
 # v3.0.0 (Jun 25, 2019)
@@ -9,6 +10,9 @@
  * BREAKING CHANGE: Plugins are installed into `"~/.appcelerator/appcd/plugins"` instead of the
    `"appcd-default-plugins/plugins"` directory to avoid permission issues when npm installing
    `appcd` globally using `sudo` due to npm dropping root before running the post install script.
+   UPDATE: Turns out that this won't work because unless there's an explicit `user` set in the
+   npm config, it defaults to the user `nobody` which does not have write permissions to the
+   user's home directory.
 
 # v2.0.0 (Jun 6, 2019)